# server/src/doc/api.yml
openapi: 3.0.0
info:
  title: PumaPay PoC API
  version: 1.0.0-alpha
  contact:
    name: PumaPay Team
    email: dev@pumapay.io
servers:
  - url: 'http://localhost:8080/api/v1'
    description: Local server
  - url: 'http://developers.pumapay.io:8080/api/v1'
    description: Test server
paths:
...
  /transaction/package/tx/metamask/{sessionID}/{packageID}:
    get:
      summary: Retrieve transaction data for credit package with session ID
      operationId: retrieveTransactionDataForPackageWithSessionID
      tags:
        - Credit Package Transaction
      parameters:
      - name: "sessionID"
        in: path
        schema:
          type: string
        description: Current session ID
        required: true
      - name: "packageID"
        in: path
        schema:
          type: string
        description: ID of credit packages details to be returned
        required: true
      security:
        - TokenAuth: []
      responses:
        '200':
          description: Succesfully returns transacation data for credit package
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
  /transaction/package/tx/wallet/plain/{userID}/{sessionID}/{packageID}:
    get:
      summary: Retrieve transaction data for credit package with session ID
      operationId: retrieveTransactionForPackageWithSessionID
      tags:
        - Credit Package Transaction
      parameters:
      - name: "userID"
        in: path
        schema:
          type: string
        description: User ID that bought the credits 
        required: true
      - name: "sessionID"
        in: path
        schema:
          type: string
        description: Current sessionID
        required: true
      - name: "packageID"
        in: path
        schema:
          type: string
        description: ID of credit packages details to be returned
        required: true
      responses:
        '200':
          description: Succesfully returns transacation data for credit package
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Transaction'
  /transaction/package/txStatus/{packageID}/{userID}/session/{sessionID}:
    get:
      summary: Retrieve transaction status for session of user ID with package ID 
      operationId: getTxStatusForSessionIDOfUserWithPackageID
      tags:
        - Credit Package Transaction
      parameters:
      - name: "sessionID"
        in: path
        schema: 
          type: string
        description: Current sessionID
        required: true
      - name: "userID"
        in: path
        schema:
          type: string
        description: ID of user for which the credits will be added
        required: true
      - name: "packageID"
        in: path
        schema:
          type: string
        description: ID of the selected credit package
        required: true
      - name: txHash
        in: query
        schema:
          type: string
        description: The txHash of the transaction
        required: false
      - name: status 
        in: query
        schema:
          type: integer
        description: The status of the transaction (initiated = -1, scanned = 0, open = 1, approved = 2, declined = 3, cancelled = 4)
        required: true
      - name: fromApp
        in: query
        schema:
          type: integer
        description: Indicates if the transaction came from the wallet app
        required: true
      security:
        - TokenAuth: []
      responses:
        '200':
          description: Succesfully returns the status of the transaction
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Session'
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
    Response:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
        message:
          type: string
        sessionID:
          type: string
        data:
          type: array
          items: {}
        errcode:
          type: string
...
    Session:
      type: object
      properties:
        sessionID:
          type: string
        txHash:
          type: string
        status:
          type: number
        fromPumaWallet:
          type: boolean
    Transaction:
      type: object
      properties:
        callback:
          type: string
        description:
          type: string
        name:
          type: string
        networkid:
          type: number
        to:
          type: string
        value:
          type: number
        signature:
          type: string
...